+++ apps/Standalone/src/templates/app/TemplatesStandaloneDesigner.tsx
@@ -18,6 +18,7 @@ import {
   StandardOperationManifestService,
   ConsumptionConnectionService,
   guid,
+  setObjectPropertyValue,
 } from '@microsoft/logic-apps-shared';
 import {
   getConnectionStandard,
@@ -357,8 +358,11 @@ const getServices = (
 
   const templateService = isConsumption
     ? new BaseTemplateService({
-        openBladeAfterCreate: (workflowName: string) => {
-          console.log('Open blade after create, workflowName is: ', workflowName);
+        openBladeAfterCreate: (workflowName: string | undefined) => {
+          window.alert(`Open blade after create, workflowName is: ${workflowName}`);
+        },
+        onAddBlankWorkflow: () => {
+          console.log('On add blank workflow click');
         },
       })
     : new StandardTemplateService({
@@ -369,8 +373,11 @@ const getServices = (
           subscription: apiVersion,
           gateway: '2018-11-01',
         },
-        openBladeAfterCreate: (workflowName: string) => {
-          console.log('Open blade after create, workflowName is: ', workflowName);
+        openBladeAfterCreate: (workflowName: string | undefined) => {
+          window.alert(`Open blade after create, workflowName is: ${workflowName}`);
+        },
+        onAddBlankWorkflow: () => {
+          console.log('On add blank workflow click');
         },
       });
 
@@ -398,6 +405,49 @@ const replaceAllStringInAllWorkflows = (workflows: StringifiedWorkflow[], oldStr
   });
 };
 
+const removeUnusedConnections = (
+  connectionsData: ConnectionsData,
+  connections: ConnectionMapping
+): { connectionsData: ConnectionsData; connections: ConnectionMapping } => {
+  const { references, mapping } = clone(connections);
+  const updatedConnectionsData = clone(connectionsData);
+  for (const connectionKey of Object.keys(mapping)) {
+    const referenceKey = mapping[connectionKey];
+    const isArmResource = isArmResourceId(references[referenceKey]?.api.id);
+
+    if (connectionKey !== referenceKey && referenceKey.startsWith(connectionKey) && !isArmResource) {
+      setObjectPropertyValue(
+        updatedConnectionsData.serviceProviderConnections ?? {},
+        [connectionKey],
+        updatedConnectionsData.serviceProviderConnections?.[referenceKey]
+      );
+
+      delete references[referenceKey];
+      delete updatedConnectionsData.serviceProviderConnections?.[referenceKey];
+
+      mapping[connectionKey] = connectionKey;
+    }
+
+    if (!isArmResource) {
+      const serviceProviderConnections = { ...(updatedConnectionsData.serviceProviderConnections ?? {}) };
+      for (const key of Object.keys(serviceProviderConnections)) {
+        if (key !== connectionKey && key.startsWith(connectionKey)) {
+          delete updatedConnectionsData.serviceProviderConnections?.[key];
+        }
+      }
+
+      const currentReferences = { ...references };
+      for (const key of Object.keys(currentReferences)) {
+        if (key !== connectionKey && key.startsWith(connectionKey)) {
+          delete references[key];
+        }
+      }
+    }
+  }
+
+  return { connectionsData: updatedConnectionsData, connections: { references, mapping } };
+};
+
 const updateConnectionsDataWithNewConnections = async (
   originalConnectionsData: ConnectionsData,
   settingProperties: Record<string, string>,
@@ -405,9 +455,11 @@ const updateConnectionsDataWithNewConnections = async (
   workflowsJsonString: StringifiedWorkflow[],
   workflowName: string
 ): Promise<{ connectionsData: ConnectionsData; settingProperties: Record<string, string>; workflowsJsonString: StringifiedWorkflow[] }> => {
-  const { references, mapping } = connections;
+  const {
+    connectionsData: updatedConnectionsData,
+    connections: { references, mapping },
+  } = removeUnusedConnections(originalConnectionsData, connections);
   let updatedSettings = { ...settingProperties };
-  const updatedConnectionsData = { ...originalConnectionsData };
   let updatedWorkflowsJsonString = workflowsJsonString;
   let updatedConnectionsJsonString = JSON.stringify(updatedConnectionsData);
   const referencesToProcess: string[] = [];