+++ libs/designer/src/lib/ui/templates/connections/workflowconnections.tsx
@@ -107,7 +107,7 @@ export const WorkflowConnections = ({ connections }: WorkflowConnectionsProps) =
     }
 
     // Sort the items.
-    const sortedItems = _copyAndSort(connectionsList(), column.fieldName as string, isSortedDescending);
+    const sortedItems = copyAndSort(connectionsList(), column.fieldName as string, isSortedDescending);
     setConnectionsList(sortedItems);
     setColumns(
       columns().map((col) => {
@@ -488,17 +488,22 @@ const ConnectionsList = ({
   return <IconButton menuIconProps={{ iconName: 'More' }} menuProps={menuProps} className="msla-template-connection-list" />;
 };
 
-const setupTemplateConnection = async (key: string, connectorId: string, connection: Connection, dispatch: AppDispatch): Promise<void> => {
+const setupTemplateConnection = async (
+  connectionKey: string,
+  connectorId: string,
+  connection: Connection,
+  dispatch: AppDispatch
+): Promise<void> => {
   await ConnectionService().setupConnectionIfNeeded(connection);
   const connector = await getConnector(connectorId);
-  dispatch(updateTemplateConnection({ connector, connection: connection, nodeId: key }));
+  dispatch(updateTemplateConnection({ connector, connection: connection, nodeId: connectionKey, connectionKey }));
 };
 
 const getConnectionDisplayName = (connection: Connection): string => {
   return connection.properties.displayName ?? connection.id.split('/').slice(-1)[0];
 };
 
-function _copyAndSort(items: ConnectionItem[], columnKey: string, isSortedDescending?: boolean): ConnectionItem[] {
+const copyAndSort = (items: ConnectionItem[], columnKey: string, isSortedDescending?: boolean): ConnectionItem[] => {
   const keyPath =
     columnKey === '$name' ? ['connectorDisplayName'] : columnKey === '$status' ? ['hasConnection'] : ['connection', 'displayName'];
   return items.slice(0).sort((a: ConnectionItem, b: ConnectionItem) => {
@@ -510,4 +515,4 @@ function _copyAndSort(items: ConnectionItem[], columnKey: string, isSortedDescen
       ? 1
       : -1;
   });
-}
+};