+++ libs/logic-apps-shared/src/designer-client-services/lib/standard/search.ts
@@ -80,7 +80,11 @@ export class StandardSearchService extends BaseSearchService {
     const isAzureConnectorsEnabled = this.options.apiHubServiceDetails.subscriptionId !== undefined;
     const filteredApiOperations = isAzureConnectorsEnabled ? response.value : filterAzureConnection(response.value);
 
-    return [...filteredApiOperations, ...getClientBuiltInOperations(filterOperation), ...(showStatefulOperations ? [ClientOperationsData.slidingWindowOperation] : [])];
+    return [
+      ...filteredApiOperations,
+      ...getClientBuiltInOperations(filterOperation),
+      ...(showStatefulOperations ? [ClientOperationsData.slidingWindowOperation] : []),
+    ];
   }
 
   public async getCustomOperationsByPage(page: number): Promise<DiscoveryOpArray> {
@@ -96,7 +100,7 @@ export class StandardSearchService extends BaseSearchService {
       const uri = `/subscriptions/${subscriptionId}/providers/Microsoft.Web/locations/${location}/apiOperations`;
       const queryParameters: QueryParameters = {
         'api-version': apiVersion,
-        $filter: `properties/trigger eq null and type eq 'Microsoft.Web/customApis/apiOperations' and ${ISE_RESOURCE_ID} eq null`,
+        $filter: `type eq 'Microsoft.Web/customApis/apiOperations' and ${ISE_RESOURCE_ID} eq null`,
       };
       // const response = await this.pagedBatchAzureResourceRequests(page, uri, queryParameters, 1);
       const { value } = await this.getAzureResourceByPage(uri, queryParameters, page, 100);